tuner class
